modularised EventsGrid into EventsCard and EventsShowcase
